Añadido cliente personalizado de FACe
- Creados StegeableTrait, FaceTrait y Faceb2bTrait
- Actualizados clientes de FACe y FACeB2B
- Añadido CustomFaceClient
- Actualizada documentación
